requirements.txt
altgraph==0.15
Appium-Python-Client==0.28
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
crypto==1.4.1
defusedxml==0.5.0
diff-match-patch==20121119
Django==2.0.4
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-formtools==2.1
django-guardian==1.4.9
django-import-export==1.0.0
django-redis==4.10.0
django-reversion==2.0.13
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
drf-extensions==0.4.0
et-xmlfile==1.0.1
future==0.16.0
get==1.0.2
httplib2==0.9.2
idna==2.6
itypes==1.1.0
jdcal==1.3
Jinja2==2.10
macholib==1.9
MarkupSafe==1.1.0
Naked==0.1.31
oauthlib==3.0.1
odfpy==1.3.6
openpyxl==2.5.1
pefile==2017.11.5
Pillow==5.1.0
post==1.0.1
public==1.0.1
publisher==0.11
pycryptodome==3.7.3
PyInstaller==3.3.1
PyJWT==1.7.1
PyMySQL==0.8.0
pypiwin32==220
python3-openid==3.1.0
pytz==2018.3
PyYAML==3.12
query-string==1.0.1
raven==6.10.0
redis==3.2.0
request==1.0.1
requests==2.18.4
requests-oauthlib==1.2.0
selenium==3.14.0
shellescape==3.4.1
six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
virtualenv==15.2.0
virtualenvwrapper-win==1.2.5
wxPython==4.0.0b2
xadmin==2.0.1
xlrd==1.1.0
xlwt==1.3.0
